+++ CHANGELOG.rst
@@ -1,10 +1,13 @@
 Changes since 0.6.0 onwards.
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-0.9.2 DATEDATEDATEDATE
+0.9.2 (2020-09-25)
 ~~~~~~~~~~~~~~~~~~
 * Removes autograd from requirements in setup.py
 * Added solar options to funcs.massradius()
+* Changed default thin=4 to thin=1 in dataset.emcee_sampler()
+* Fixed bug in Multivisit for default log_Q value (#155)
+* Added PlanetProperties
 
 0.9.1 (2020-09-10)
 ~~~~~~~~~~~~~~~~~~